+++ R/full-join.R
@@ -27,3 +27,12 @@ full_join.treedata <- function(x, y, by = NULL,
     }
     return(x)
 }
+
+##' @method full_join phylo
+##' @export
+full_join.phylo <- function(x, y, by = NULL,
+                            copy = FALSE, suffix = c(".x", ".y"), ...) {
+    full_join(as_tibble(x), y = y, by = by,
+              copy = copy, suffix = suffix, ...) %>%
+        as.treedata
+}